168: Support search pagination (#174)
